pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "next3d"
version = "0.1.0"
description = "Semantic 3D geometry cognition system — structured, queryable, AI-native CAD understanding"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "OCP-stubs>=7.7",
    "cadquery>=2.4",
    "networkx>=3.2",
    "pydantic>=2.5",
    "click>=8.1",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
]

[project.scripts]
next3d = "next3d.cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/next3d"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
